pyproject.toml
[tool.poetry]
name = "swarmsort"
version = "0.1.0"
description = "Standalone SwarmSort multi-object tracker with deep learning embeddings"
authors = ["Charles Fosseprez <charles.fosseprez.pro@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/your-org/swarmsort"
repository = "https://github.com/your-org/swarmsort"
keywords = ["tracking", "computer-vision", "multi-object-tracking", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
numpy = "^1.24.0"
scipy = "^1.10.0"
numba = "^0.57.0"
loguru = "^0.7.0"
matplotlib = {version = "^3.7.0", optional = true}
pyyaml = "^6.0"
opencv-python = "^4.8.0"
cupy-cuda11x = {version = "^12.0.0", optional = true}
cupy-cuda12x = {version = "^12.0.0", optional = true}

[tool.poetry.extras]
plotting = ["matplotlib"]
gpu11 = ["cupy-cuda11x"]
gpu = ["cupy-cuda12x"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=swarmsort --cov-report=term-missing --cov-report=html -v"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-benchmark = "^4.0.0"
pytest-timeout = "^2.2.0"
pytest-mock = "^3.11.0"
pytest-xdist = "^3.3.0"
black = {extras = ["jupyter"], version = "^23.9.0"}
mypy = "^1.5.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
bandit = "^1.7.5"
safety = "^2.3.5"
psutil = "^5.9.5"
